[scudo] [minesweeper] Track active regions
Add mechanism to track which regions of memory are active,
i.e. either have allocated blocks or have not been released since the last deallocation.
Granularity is per page for primary and per block for secondary.

Add mechanism to iterate through these regions.
This is to allow minesweeper to scan active memory.